child layout is in charge of composing
This allows subclassing Layout in order to control whether parent
layouts will be composed into this layout. For example, if a child
layout is able to determine that it or its parents are the same as
what the browser already has (which takes cooperation from frontend
scripting) and transparently avoids adding additional unnecessary
layouts to the response.